reorg controllers in prep for number 3
