+++ lib/mars_base_10/controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative "controller/graph_rover"
+require_relative "controller/group_room"
+
+require_relative 'ship'
+require_relative 'stack'
+require_relative 'subject'
+
+module Controller
+end
